Split out code and replay working .. yay
