--- src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegmentTypes.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 0ebd5809d41d4029bcf48aeb15c5c551
-timeCreated: 1716914937
