+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 467fee0ce5364d04b2ce3d05391377c8
+timeCreated: 1717514587
