+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/JsonConverters.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3c4aefb0bae24e01aaecda590931f8fc
+timeCreated: 1717514236
