+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AndKeyFrameCriteriaData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e8470d6d267b4ac5bce84d700e59db05
+timeCreated: 1717514966
