+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AndOr.cs
@@ -1,0 +1,8 @@
+﻿namespace StateRecorder.BotSegments.Models
+{
+    public enum AndOr
+    {
+        And,
+        Or
+    }
+}
