+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotAction.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Text;
+using Newtonsoft.Json;
+using RegressionGames.StateRecorder.JsonConverters;
+using StateRecorder.BotSegments.JsonConverters;
+
+namespace StateRecorder.BotSegments.Models
+{
+    [JsonConverter(typeof(BotActionJsonConverter))]
+    [Serializable]
+    public class BotAction
+    {
+        public BotActionType type;
+        public IBotActionData data;
+        public bool IsCompleted => data.IsCompleted();
+
+        public void ReplayReset()
+        {
+            data.ReplayReset();
+        }
+
+        public void WriteToStringBuilder(StringBuilder stringBuilder)
+        {
+            stringBuilder.Append("{\"type\":");
+            StringJsonConverter.WriteToStringBuilder(stringBuilder, type.ToString());
+            stringBuilder.Append(",\"data\":");
+            data.WriteToStringBuilder(stringBuilder);
+            stringBuilder.Append("}");
+        }
+    }
+}
