+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/BotActionType.cs
@@ -1,0 +1,21 @@
+﻿namespace StateRecorder.BotSegments.Models
+{
+    public enum BotActionType
+    {
+        None,
+        InputPlayback,
+
+        //FUTURE
+        //Timer,
+        //FrameCount,
+        //Behaviour,
+        //AgentBuilder,
+        //OrchestratedInput,
+        //RandomMouse_ClickPixel,
+        //RandomMouse_ClickObject,
+        //RandomKeyboard_Key,
+        //RandomKeyboard_ActionableKey,
+        //RandomJoystick_Axis1|Axis2|Axis3,Key
+
+    }
+}
