+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/CountRule.cs
@@ -1,0 +1,10 @@
+﻿namespace StateRecorder.BotSegments.Models
+{
+    public enum CountRule
+    {
+        Zero,
+        NonZero,
+        GreaterThanEqual,
+        LessThanEqual
+    }
+}
