+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/CountRule.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e6971dee00324c4296e46080ee267cad
+timeCreated: 1717515067
