+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/IBotActionData.cs
@@ -1,0 +1,13 @@
+﻿using System.Text;
+
+namespace StateRecorder.BotSegments.Models
+{
+    public interface IBotActionData
+    {
+        public void WriteToStringBuilder(StringBuilder stringBuilder);
+
+        public void ReplayReset();
+
+        public bool IsCompleted();
+    }
+}
