+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/IBotActionData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d0e72d0578dd428193ef08462bf89ce7
+timeCreated: 1717514406
