+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/IKeyFrameCriteriaData.cs
@@ -1,0 +1,9 @@
+﻿using System.Text;
+
+namespace StateRecorder.BotSegments.Models
+{
+    public interface IKeyFrameCriteriaData
+    {
+        public void WriteToStringBuilder(StringBuilder stringBuilder);
+    }
+}
