+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/IKeyFrameCriteriaData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 09436a1a86ce4e9d8d9b73815bff2a1c
+timeCreated: 1717514464
