+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/KeyFrameCriteria.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 43004bae4c2241079a338ff8554dc3f8
+timeCreated: 1717514491
