+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/KeyFrameCriteriaType.cs
@@ -1,0 +1,15 @@
+﻿namespace StateRecorder.BotSegments.Models
+{
+    public enum KeyFrameCriteriaType
+    {
+        Or,
+        And,
+        UIPixelHash,
+        NormalizedPath,
+
+        //FUTURE
+        //Path,
+        //XPath,
+        //BehaviourComplete
+    }
+}
