+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/KeyFrameCriteriaType.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9c59d81bf9fe40b588a8f5d8a257b9ed
+timeCreated: 1717515112
