+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/PathKeyFrameCriteriaData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f459f982675e42dea58e568aeef02da8
+timeCreated: 1717514939
