+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/NormalizedPathCriteriaEvaluator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2a459d9ef11e4dc8af2377b6a4ab1a68
+timeCreated: 1717514700
