+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/InputData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: eb080496c94c480d83dbaea728e242f6
+timeCreated: 1717516360
