+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/TransformStatus.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 06bb64ee305d46bfbeefeec4dad0da6e
+timeCreated: 1717516564
