--- src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/StateElementDeltaType.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 837cfe2dffb740dea9b00ec1d98290bf
-timeCreated: 1716216799
